# backend/src/main/resources/META-INF/openapi.yaml
openapi: 3.1.0
info:
  title: Scaramutti TMS API
  version: 1.0.0
  contact:
    name: Equipo Scaramutti TMS
    email: ascaramutti@hotmail.com
  description: |
    API del sistema de gestión Scaramutti TMS. Esta es la especificación
    **runtime** que refleja lo efectivamente implementado en el backend.
    El contrato completo de diseño (incluye módulos aún no implementados)
    vive en `api/openapi.yaml` en la raíz del repo.

    ## Convenciones generales

    - **Versionado**: en URL `/api/v1/...`. Cambios breaking suben mayor (`v2`).
    - **Autenticación**: JWT en header `Authorization: Bearer <token>`. Todos los
      endpoints requieren auth excepto los marcados con `security: []`
      (`/auth/login`, `/auth/refresh`).
    - **Errores**: formato Problem Details (RFC 7807). Incluyen `traceId` para
      correlacionar con logs. Errores de validación incluyen array `errors` por campo.
    - **Idioma**: español (es-PE). Las respuestas no son multi-idioma.
    - **Timezone**: todos los timestamps en UTC, formato ISO 8601 con offset `Z`.
    - **Naming**: snake_case en BD, camelCase en JSON.

    ## Roles

    Siete roles del sistema: `admin`, `sales`, `dispatcher`, `general_manager`, `operations_manager`, `finance_manager`, `warehouse_keeper`.
servers:
  - url: http://localhost:8080/api/v1
    description: Local (dev). El URL real se inyecta en cada cliente vía variable de entorno.
security:
  - bearerAuth: []
...
paths:
...
  /warehouse/suppliers:
    get:
      tags: [Warehouse]
      operationId: listWarehouseSuppliers
      summary: Buscar proveedores (paginado)
      description: |
        Con `q` filtra por substring case-insensitive en `name` y `ruc` (ILIKE
        acelerado por GIN trgm `idx_suppliers_name_trgm`), ranking `similarity`
        DESC, tiebreak `name` ASC. Multi-palabra: `q` se tokeniza por espacios y
        cada palabra debe matchear en ALGÚN campo (AND de ORs) — RN-WH14.
        Sin `q`: `name` ASC. Para no filtrar, OMITIR el parámetro (no enviarlo
        vacío).
      x-required-roles: [admin, general_manager, operations_manager, finance_manager, warehouse_keeper]
      parameters:
        - { name: q, in: query, schema: { type: string, minLength: 3, maxLength: 200 } }
        - { name: isActive, in: query, schema: { type: boolean } }
        - { name: page, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200':
          description: Página de proveedores
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageOfWarehouseSupplier' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Warehouse]
      operationId: createWarehouseSupplier
      summary: Crear proveedor (al vuelo desde el form de factura)
      description: '`name` y `ruc` únicos → `409 WH-010` (el `detail` dice cuál chocó).'
      x-required-roles: [admin, general_manager, operations_manager, finance_manager, warehouse_keeper]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WarehouseSupplierRequest' }
            example: { name: 'REPUESTOS DIÉSEL S.A.C.', ruc: '20512345678', phone: '014567890', contactName: 'Juan Ramos' }
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WarehouseSupplierResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/ConflictDuplicate' }
  /warehouse/products:
    get:
      tags: [Warehouse]
      operationId: listWarehouseProducts
      summary: Listar productos CON su stock actual (Existencias y comboboxes)
      description: |
        Un solo listado sirve a la pantalla Existencias y a los comboboxes de
        producto: cada fila trae `stock` y `lowStock` LEÍDOS de la VIEW
        `almacen.product_stock` (RN-WH1, nunca persistidos). `lowStock = stock <
        minStock` (estricto) se define UNA vez en la VIEW (RN-WH11).
        `q` (≥ 3) filtra por substring case-insensitive en `name`, `code`,
        `brand` y `partNumber`; multi-palabra: cada palabra debe matchear en
        ALGÚN campo (AND de ORs, RN-WH14), ranking `similarity` DESC, tiebreak
        `name` ASC. Sin `q`: `name` ASC. `lowOnly=true` restringe a los productos
        bajo umbral. Para no filtrar, OMITIR el parámetro.
      x-required-roles: [admin, general_manager, operations_manager, finance_manager, warehouse_keeper]
      parameters:
        - { name: q, in: query, schema: { type: string, minLength: 3, maxLength: 200 } }
        - { name: categoryId, in: query, schema: { type: integer } }
        - { name: isActive, in: query, schema: { type: boolean } }
        - name: lowOnly
          in: query
          description: 'Si `true`, solo productos con `stock < minStock` (badge "Bajo").'
          schema: { type: boolean, default: false }
        - { name: page, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200':
          description: Página de productos con stock
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageOfWarehouseProduct' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Warehouse]
      operationId: createWarehouseProduct
      summary: Crear producto (al vuelo desde la factura o desde Existencias)
      description: |
        Identidad ÚNICA compuesta `(name, brand, partNumber)` → `409 WH-010`
        (Δ-2, RN-WH10): mismo nombre y marca pueden coexistir con distinto n° de
        parte (variantes/modelos); comparación case-insensitive y los opcionales
        vacíos cuentan como "sin valor" (el backend normaliza trim + vacío→null;
        índice con COALESCE — dos genéricos idénticos sin marca/parte también
        chocan). El `code` (SKU) lo asigna siempre el backend (`PRO-NNNN`,
        correlativo); no se acepta en el request.
        `attributes` = objeto JSON libre clave-valor (JSONB).
        Un producto nuevo **nace con stock 0** — su primera factura es su primer
        movimiento; el corte inicial es SOLO para inventario pre-existente (RN-WH6).
        Duplicados: el guardián es el índice único de identidad compuesta → `409
        WH-010`; no hay cabecera de idempotencia (redundante frente al índice,
        consistente con el resto del módulo).
      x-required-roles: [admin, general_manager, operations_manager, finance_manager, warehouse_keeper]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WarehouseProductRequest' }
            example:
              name: 'Filtro de aceite XYZ'
              categoryId: 7
              unitOfMeasureId: 1
              brand: 'Bosch'
              partNumber: 'F026407123'
              attributes: { rosca: '3/4-16' }
              minStock: 4
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WarehouseProductResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/ConflictDuplicate' }
  /warehouse/products/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      tags: [Warehouse]
      operationId: getWarehouseProduct
      summary: Obtener producto por ID (con stock actual)
      description: |
        El header `ETag` (versión opaca del `updatedAt`) es el que exige el PUT
        vía `If-Match`. Usar el HEADER, no el `updatedAt` del body.
      x-required-roles: [admin, general_manager, operations_manager, finance_manager, warehouse_keeper]
      responses:
        '200':
          description: Producto con stock
          headers:
            ETag: { schema: { type: string }, description: Versión opaca para If-Match }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WarehouseProductResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Warehouse]
      operationId: updateWarehouseProduct
      summary: Actualizar producto (catálogo — sin justificación)
      description: |
        Edición de catálogo: nombre, categoría, marca, nº de parte, atributos,
        `minStock`, observaciones, `isActive` (soft delete). La **unidad de
        medida NO viaja** — se fija al crear y es inmutable (P-1). NO exige
        justificación — la regla editar-con-justificación (RN-WH4) aplica a
        movimientos de stock, no al catálogo (RN-WH15). Identidad compuesta
        `(name, brand, partNumber)` única (Δ-2) y `code` único → `409 WH-010`.
        No afecta el kardex (los movimientos referencian por id).
        `If-Match` obligatorio (optimistic locking): versión vieja → `412
        COM-004`. El ETag a enviar es el HEADER del GET o del PUT anterior.
      x-required-roles: [admin, general_manager, operations_manager, finance_manager, warehouse_keeper]
      parameters:
        - { name: If-Match, in: header, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WarehouseProductUpdateRequest' }
      responses:
        '200':
          description: Actualizado
          headers:
            ETag: { schema: { type: string }, description: Nueva versión opaca para el próximo If-Match }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WarehouseProductResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/ConflictDuplicate' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
  /warehouse/products/{id}/stock:
    get:
      tags: [Warehouse]
      operationId: getWarehouseProductStock
      summary: Stock disponible en vivo (form de retiro)
      description: |
        Lectura puntual de la VIEW `almacen.product_stock` — el form de retiro
        muestra el disponible al elegir producto y bloquea el submit si excede
        (espejo UX). La validación AUTORITATIVA sigue siendo la del POST/PUT del
        retiro en transacción (`409 WH-001`, RN-WH2).
      x-required-roles: [admin, general_manager, operations_manager, finance_manager, warehouse_keeper]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Stock del producto
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WarehouseProductStockResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /warehouse/products/{id}/kardex:
    get:
      tags: [Warehouse]
      operationId: getWarehouseProductKardex
      summary: Kardex del producto (movimientos con saldo corrido)
      description: |
        Lee la VIEW `almacen.stock_movements`: APERTURA ∪ ENTRADAS activas ∪
        SALIDAS activas (los ANULADOS no aparecen). Orden `movedAt` DESC.
        `balance` = saldo DESPUÉS del movimiento, calculado por el backend sobre
        la historia completa: correcto aunque se pagine o filtre por fechas
        (RN-WH13). `quantity` siempre positiva (el signo lo da `movementType`).
      x-required-roles: [admin, general_manager, operations_manager, finance_manager, warehouse_keeper]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: dateFrom, in: query, description: 'Desde (inclusive, América/Lima)', schema: { type: string, format: date } }
        - { name: dateTo, in: query, description: 'Hasta (inclusive del día completo, América/Lima)', schema: { type: string, format: date } }
        - { name: page, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200':
          description: Página de movimientos
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageOfWarehouseKardexMovement' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT firmado con RS256. Claims esperados:
        `sub` (userId), `username`, `role`, `fullName`, `position`, `type`, `iat`, `exp`.
...
  responses:
    Unauthorized:
      description: Token de acceso ausente, expirado o inválido
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            tokenMissing:
              summary: No se envió header Authorization
              value:
                type: 'urn:tms:error:auth-006'
                title: 'Token missing'
                status: 401
                detail: 'Falta el header Authorization con un token Bearer'
                instance: '/api/v1/auth/me'
                code: 'AUTH-006'
                traceId: '4a5b6c7d-8e9f-0a1b-2c3d-4e5f6a7b8c9d'
            tokenExpired:
              summary: Token de acceso expirado
              value:
                type: 'urn:tms:error:auth-007'
                title: 'Token expired'
                status: 401
                detail: 'El token de acceso expiró. Use /auth/refresh para renovarlo.'
                instance: '/api/v1/auth/me'
                code: 'AUTH-007'
                traceId: '8d2e4f6a-1b3c-4d5e-7f8a-9b0c1d2e3f4a'
            tokenInvalid:
              summary: Token mal formado o no firmado por este servidor
              value:
                type: 'urn:tms:error:auth-008'
                title: 'Token invalid'
                status: 401
                detail: 'El token de acceso es inválido o está mal formado'
                instance: '/api/v1/auth/me'
                code: 'AUTH-008'
                traceId: '5c1a2b3d-4e5f-6a7b-8c9d-0e1f2a3b4c5d'
    BadRequest:
      description: Solicitud inválida (validación, formato, valores fuera de rango)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            validationFailed:
              summary: Errores de validación por campo
              value:
                type: 'urn:tms:error:com-001'
                title: 'Validation failed'
                status: 400
                detail: 'La solicitud contiene errores de validación'
                instance: '/api/v1/clients'
                code: 'COM-001'
                traceId: '7c4d2f8a-1b3e-4f9a-8c2d-5e7b1a9c3d4f'
                errors:
                  - field: 'name'
                    message: 'no debe estar vacío'
                    code: 'REQUIRED'
                  - field: 'ruc'
                    message: 'debe tener exactamente 11 dígitos'
                    code: 'PATTERN'
    Forbidden:
      description: Autenticado pero sin permisos para esta operación
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            insufficientRole:
              summary: Rol insuficiente para la operación
              value:
                type: 'urn:tms:error:com-003'
                title: 'Forbidden'
                status: 403
                detail: 'No tiene permisos para acceder a este recurso'
                instance: '/api/v1/clients'
                code: 'COM-003'
                traceId: '6f4e3d2c-1b0a-9f8e-7d6c-5b4a3f2e1d0c'
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            quotationNotFound:
              summary: Cotización inexistente
              value:
                type: 'urn:tms:error:quo-003'
                title: 'Resource not found'
                status: 404
                detail: 'La cotización con id 999 no existe'
                instance: '/api/v1/quotations/999'
                code: 'QUO-003'
                traceId: '2b3c4d5e-6f7a-8b9c-0d1e-2f3a4b5c6d7e'
            productNotFound:
              summary: Producto de almacén inexistente
              value:
                type: 'urn:tms:error:wh-003'
                title: 'Resource not found'
                status: 404
                detail: 'El producto con id 999 no existe'
                instance: '/api/v1/warehouse/products/999'
                code: 'WH-003'
                traceId: '2b3c4d5e-6f7a-8b9c-0d1e-2f3a4b5c6d7e'
    PreconditionFailed:
      description: Optimistic lock — el recurso fue modificado desde que se leyó
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            staleVersion:
              summary: Versión obsoleta (otro usuario editó primero)
              value:
                type: 'urn:tms:error:com-004'
                title: 'Precondition failed'
                status: 412
                detail: 'El recurso fue modificado por otro usuario. Recargue antes de editar.'
                instance: '/api/v1/quotations/142'
                code: 'COM-004'
                traceId: '0a1b2c3d-4e5f-6a7b-8c9d-0e1f2a3b4c5d'
    ConflictDuplicate:
      description: Duplicado de catálogo del módulo Almacén — nombre/código/RUC ya existe → WH-010
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'urn:tms:error:wh-010'
            title: 'Conflict'
            status: 409
            detail: 'Ya existe un registro con ese valor en el catálogo'
            instance: '/api/v1/warehouse/suppliers'
            code: 'WH-010'
            traceId: 'a0b1c2d3-4e5f-6a7b-8c9d-0e1f2a3b4c5d'
  schemas:
...
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/UserRef'
        - type: object
          required: [role, isActive]
          properties:
            role:     { $ref: '#/components/schemas/UserRole' }
            isActive: { type: boolean }

    # -------- Catálogos: Currency ------------------------------------------
...
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages, numberOfElements, first, last, empty]
      properties:
        page:              { type: integer, format: int32, example: 0 }
        size:              { type: integer, format: int32, example: 20 }
        totalElements:     { type: integer, format: int64, example: 137 }
        totalPages:        { type: integer, format: int32, example: 7 }
        numberOfElements:  { type: integer, format: int32, description: 'Cantidad de items en la página actual', example: 20 }
        first:             { type: boolean, description: '¿Es la primera página?' }
        last:              { type: boolean, description: '¿Es la última página?' }
        empty:             { type: boolean, description: '¿La página actual está vacía?' }
...
    WarehouseSupplierResponse:
      type: object
      required: [id, name, isActive, createdAt]
      properties:
        id:          { type: integer }
        name:        { type: string, maxLength: 200, example: 'REPUESTOS DIÉSEL S.A.C.' }
        ruc:         { type: [string, 'null'], pattern: '^\d{11}$' }
        phone:       { type: [string, 'null'], pattern: '^\d{9}$' }
        contactName: { type: [string, 'null'], maxLength: 100 }
        isActive:    { type: boolean }
        createdAt:   { type: string, format: date-time }
    WarehouseSupplierRequest:
      type: object
      required: [name]
      properties:
        name:        { type: string, minLength: 3, maxLength: 200 }
        ruc:         { type: [string, 'null'], pattern: '^\d{11}$' }
        phone:       { type: [string, 'null'], pattern: '^\d{9}$' }
        contactName: { type: [string, 'null'], maxLength: 100 }
    PageOfWarehouseSupplier:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/WarehouseSupplierResponse' }
    PageOfWarehouseProduct:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/WarehouseProductResponse' }
    WarehouseProductRequest:
      type: object
      required: [name, categoryId, unitOfMeasureId]
      properties:
        name:            { type: string, minLength: 3, maxLength: 200, description: 'Único COMPUESTO con brand+partNumber (Δ-2, RN-WH10)' }
        categoryId:      { type: integer }
        unitOfMeasureId: { type: integer }
        brand:           { type: [string, 'null'], maxLength: 100 }
        partNumber:      { type: [string, 'null'], maxLength: 100 }
        attributes:
          type: object
          additionalProperties: { type: string }
          default: {}
          description: 'Características flexibles clave-valor (JSONB). Objeto libre; la UI lo edita como filas clave-valor.'
          example: { rosca: '3/4-16', viscosidad: '15W-40' }
        minStock:     { type: number, minimum: 0, default: 0, description: 'Umbral de reposición: stock < minStock → badge "Bajo" (RN-WH11)' }
        observations: { type: [string, 'null'] }
        isActive:     { type: boolean, default: true, description: 'Soft delete del catálogo. En el POST se ignora (siempre nace activo).' }
    WarehouseProductUpdateRequest:
      type: object
      description: >-
        Como WarehouseProductRequest pero SIN unitOfMeasureId: la unidad de
        medida se fija al crear y es inmutable (P-1). Su edición se evaluará más
        adelante.
      required: [name, categoryId]
      properties:
        name:            { type: string, minLength: 3, maxLength: 200, description: 'Único COMPUESTO con brand+partNumber (Δ-2, RN-WH10)' }
        categoryId:      { type: integer }
        brand:           { type: [string, 'null'], maxLength: 100 }
        partNumber:      { type: [string, 'null'], maxLength: 100 }
        attributes:
          type: object
          additionalProperties: { type: string }
          default: {}
          description: 'Características flexibles clave-valor (JSONB). Objeto libre; la UI lo edita como filas clave-valor.'
        minStock:     { type: number, minimum: 0, default: 0 }
        observations: { type: [string, 'null'] }
        isActive:     { type: boolean, default: true, description: 'Soft delete del catálogo.' }
    WarehouseProductResponse:
      type: object
      required: [id, code, name, category, unitOfMeasure, attributes, minStock, isActive, stock, lowStock, createdBy, createdAt, updatedAt]
      properties:
        id:   { type: integer }
        code: { type: string, maxLength: 30, description: 'SKU (autogenerado por el backend)' }
        name: { type: string, maxLength: 200 }
        category:
          type: object
          required: [id, name]
          properties:
            id:   { type: integer }
            name: { type: string }
        unitOfMeasure:
          type: object
          required: [id, code, name]
          properties:
            id:   { type: integer }
            code: { type: string }
            name: { type: string }
        brand:      { type: [string, 'null'], maxLength: 100 }
        partNumber: { type: [string, 'null'], maxLength: 100 }
        attributes:
          type: object
          additionalProperties: { type: string }
          description: 'Objeto JSON libre clave-valor (JSONB). Nunca `null` (default `{}`).'
        minStock:     { type: number, minimum: 0 }
        observations: { type: [string, 'null'] }
        isActive:     { type: boolean }
        stock:        { type: number, description: 'LEÍDO de la VIEW `almacen.product_stock` (RN-WH1 — nunca persistido, siempre exacto)' }
        lowStock:     { type: boolean, description: 'Derivado en código: `stock < minStock` (estricto). NO es columna.' }
        createdBy:    { $ref: '#/components/schemas/UserResponse' }
        createdAt:    { type: string, format: date-time }
        updatedAt:
          type: string
          format: date-time
          description: >-
            Versión del recurso (source del optimistic locking). Para el
            `If-Match` del PUT usar el header `ETag` del GET/PUT, NO este campo
            del body (difieren por µs entre memoria y BD → 412 espurio).
    WarehouseProductStockResponse:
      type: object
      description: Lectura puntual del stock disponible (form de retiro, validación en vivo).
      required: [productId, stock, minStock, lowStock]
      properties:
        productId: { type: integer }
        stock:     { type: number, description: 'De la VIEW `almacen.product_stock`' }
        minStock:  { type: number }
        lowStock:  { type: boolean, description: 'Derivado: `stock < minStock`' }
...
    PageOfWarehouseKardexMovement:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/WarehouseKardexMovementResponse' }
...